// src/containers/ProfileSettingsForm/types.tsx
export interface FormValues {
    username: string;
    password: string;
    email: string;
    phone: string;
}

export enum Fields {
    USERNAME = "username",
    PASSWORD = "password",
    EMAIL = "email",
    PHONE = "phone",
}

export enum PLACEHOLDERS {
    USERNAME = "Имя пользователя",
    EMAIL = "Email адресс",
    PHONE = "Номер телефона",
    PASSWORD = "Пароль",
}